# Optimize some local functions