--- a/bytecomp/lambda.ml
+++ b/bytecomp/lambda.ml
@@ -758,8 +758,8 @@ let shallow_map f = function
         ap_inlined;
         ap_specialised;
       }
-  | Lfunction { kind; params; body; attr; loc; } ->
-      Lfunction { kind; params; body = f body; attr; loc; }
+  | Lfunction { kind; params; return; body; attr; loc; } ->
+      Lfunction { kind; params; return; body = f body; attr; loc; }
   | Llet (str, k, v, e1, e2) ->
       Llet (str, k, v, f e1, f e2)
   | Lletrec (idel, e2) ->