--- a/bytecomp/lambda.mli
+++ b/bytecomp/lambda.mli
@@ -333,9 +333,15 @@ val name_lambda: let_kind -> lambda -> (Ident.t -> lambda) -> lambda
 val name_lambda_list: lambda list -> (lambda list -> lambda) -> lambda
 
 val iter_head_constructor: (lambda -> unit) -> lambda -> unit
-(** [iter_head_constructor f lam] apply [f] to only the first level of
+(** [iter_head_constructor f lam] apply [f] or [tail] to only the first level of
     sub expressions of [lam]. It does not recursively traverse the
-    expression. *)
+    expression.
+*)
+
+val shallow_iter: tail:(lambda -> unit) -> non_tail:(lambda -> unit) -> lambda -> unit
+(** Same as [iter_head_constructor], but use a different callback for sub-terms which
+    are in tail position or not. *)
+
 
 val free_variables: lambda -> Ident.Set.t
 
@@ -367,9 +373,6 @@ val map : (lambda -> lambda) -> lambda -> lambda
 
 val shallow_map  : (lambda -> lambda) -> lambda -> lambda
   (** Rewrite each immediate sub-term with the function. *)
-
-val iter_tail : (lambda -> unit) -> lambda -> unit
-  (** Apply the callback to each immediate sub-term in tail position. *)
 
 val bind : let_kind -> Ident.t -> lambda -> lambda -> lambda
 val bind_with_value_kind: