… bytecomp/translattribute.ml
… testsuite/tests/basic/localfunction.ml
… testsuite/tests/basic/localfunction.reference
… testsuite/tests/basic/ocamltests
… testsuite/tests/functors/functors.compilers.reference
… testsuite/tests/warnings/w47_inline.compilers.reference
… testsuite/tests/warnings/w47_inline.ml