This is a revised version of [MPR#6242](https://caml.inria.fr/mantis/view.php?id=6242).

[EDIT: modified description to match the latest version of the PR]

A new optimization pass is added in Simplif to avoid the overhead of allocating and calling local closures.  Functions bound to an identifier are simplified away if all references to the identifier are full applications of the function (i.e. same number of arguments) and if all these calls are in the same "tail scope".  In that case, a "staticcatch" handler is inserted around that scope and each call is turned into a "staticraise" jumping to that handler.  A special case is when a function is used only once (for a full application), which obviously fall into the case above; in that case, an existing pass further simplifies the static exception by inlining the handler (i.e. the function's body) on the call site.

The case of local functions used in tail position corresponds to a very frequent use case where multiple branches (of pattern matching or conditionals) need to finish with the same computation.  There is currently no very good way to express this pattern: (1) a local function allocates the closure (even when unused), calling the function has a cost, and this breaks optimizations restricted to a single function body (unboxing of boxed number, register allocation);  (2) a `try...with` is costly (even with raise_notrace), breaks the tail-context (i.e. can turn tail calls into non-tail ones) and is syntactically not very pleasant.

It could be the case that some cases are or will be optimized with flambda (I did not check), but (1) this would not benefit to Javascript backends (js_of_ocaml & Bucklescript) and (2) flambda is still not the default (and not very widely used I think).

Playing again the micro-benchmark from MPR6242, I get:

Default settings for ocamlopt (no flambda):
   before: 4.0s
   after: 3.3s  (18% faster)

With -inline 100:
   before: 4.0s
   after: 1.5s (62% faster)


Another micro-benchmark:
````ocaml
let foo x =
  let r = ref 0. in
  if x < 10 then r := !r +. 1.;
  let ret n = !r *. n in
  if x mod 3 = 0 then ret 2.
  else if x mod 3 = 1 then ret 1.
  else 10.

let () =
  let r = ref 0. in
  for i = 1 to 1000000000 do
    r := !r +. foo i
  done;
  Printf.printf "%f\n%!" !r
````

Default settings for ocamlopt: 
   before: 5.9s
   after: 4.2s  (29% faster)

With -inline 100:
   before: 5.9s
   after: 2.4s (59% faster)