@alainfrisch 2018-11-09
You're right that one should be careful with quadratic behavior (even though the constant factor should be very small here).  One possibility, which might be equivalent to your proposal, would be to traverse the term, keeping track during the descent of a "non-tailness-level" (incremented each time we navigate to a sub term in non-tail position).  The traversal would store the maximum level for each identifier, which allows, on the way up, to know if a given identifier is only used in tail-call position.