@gasche 2018-11-12
(The commit messages says "pseudi linear-time" instead of "pseudo", I think.)