@alainfrisch 2018-11-12
@gasche : thanks for the live review 😄 

The latest version is a generalization of the two cases initially described: it turns a function into a static handler as soon as all uses of the function are full applications, all of them in them "tail scope".  This cover in particular the case of a function used only once (including for module-level function bindings, i.e. a function introduced in a module but not exported and used only once in the module).

Nested functions are supported as well, for instance all local function bindings are eliminated in:
````ocaml
let foo x y =
  let f x = x * x * x in
  let ret n = f (x + n * y) in
  let a =
    if x mod 2 = 0 then ret 3
    else if y mod 2 = 0 then ret 4
    else f 10
  in
  a + 1
````