@alainfrisch 2018-11-12
> Misc.protect_refs [R(current_scope, scope)] (fun () -> tail lam)

True, but this might be a bit performance sensitive (called for each non-tail lambda), so I felt that the extra 3 lines of code where worth it.