@alainfrisch 2018-11-12
> (The commit messages says "pseudi linear-time" instead of "pseudo", I think.)

Yes, of course.  If the PR is accepted, it should be squashed I think, so I won't bother rewriting the history to fix this typo.