@mshinwell 2018-11-12
@lpw25 and I discussed this proposal; here are some thoughts.  (This was a comment on your first version.)

As background: one of the main design principles behind the next version of Flambda is to try to ensure that, even in the presence of high optimisation, users are able to predict and understand what the compiler is doing.  The current compiler is good in this regard and we would like not to compromise it.  The principle extends, in particular, to avoiding "cliff edges" where optimisations suddenly do (or do not) kick in and cause sudden performance changes.

One place where "cliff edges" can occur is when using optimisations that depend on the number of uses of some particular thing, such as the first optimisation proposed here: adding a second use of a local function could suddenly cause a non-negligible change in performance.  As such we think we should be a bit careful here both in terms of the specification of the optimisation, and also in its implementation.  In terms of the specification, it should be _syntactically obvious_ when the optimisation is going to take effect, so we can give a clear criterion in the manual that is easily checked by the programmer.  In terms of the implementation, this probably means doing the analysis prior to `Lambda`, marking the sub-terms appropriately based on syntactic criteria isomorphic to the equivalents on the source text of the program, then doing the optimisation later based on such markers.

Writing high-performance code sometimes requires manual guidance to be given to the compiler. 
 In this case, we should provide attributes (perhaps re-using, or at least respecting the existing inlining ones) to disable the optimisation, on both the function declaration and its use sites.

It's worth noting that these optimisations, perhaps surprisingly, are not always a win.  For example, for the tail optimisation, if the function is small it might be better just to inline it at each of the use sites (especially if the inlined versions are then variously simplified by using information about the function's arguments).  Another problem is interaction with inlining of the outer function: if the outer function gets larger by virtue of containing the code of the inner local function as a result of these optimisations, it might prevent the outer function getting inlined, which could be worse.  These kind of interactions, along with the ability if we do these optimisations in Flambda instead of at the Lambda level to integrate their results into the inlining reports, tend to support an approach where the relevant sub-terms are marked for later analysis.  We could just disable the transformation on Lambda and do it later, somehow integrating these optimisations into the usual inlining benefit calculations, or perhaps having them as a separate phase after inlining.  (We haven't spent enough time to really get to grips with how this would work yet, but these are some preliminary ideas.)

That said, we also suspect that the application of the tail optimisation prior to Flambda (but not the used-only-once optimisation) might not compromise optimisation in Flambda 2.0, because its whole _modus operandi_ is that of static catch handlers (continuations).

Another issue relates to closed local functions, which will be lifted out, so there won't be any closure allocation.  There still will be register shuffling to match the calling convention---but on the other hand, so will there probably be under these optimisations, where the shared continuation has its own calling convention assigned by the register allocator.  As such, we think it would be better to phrase both optimisations in terms of guaranteeing lack of closure allocations: "a local function syntactically used only once, or syntactically used only in tail position, is guaranteed not to allocate a closure".  Then the compiler can choose whether to lift the function out or apply one of these optimisations.

Given that these optimisations are effectively stand-alone, I also recommend they should be in a separate module, rather than lumping them together into `Simplif`.